requirements.txt
Django==5.0.6
django_ckeditor==6.7.1
